art, music, code updates
picked up some public domain music and added it to the code, same with a few pictures, although I'm trying to go more pixel art on this, I might just pixelate the images.  have been working on walking animation images and integration into the code, it's getting there.  changed the buttons on pregameui, added a mouseover effect to them as well, still a few changes to be made to fully have the game migrated over from gohomeless, but it's starting to shape up a bit.