displaying played time in newgame/loadgame screen
separate functions for loading and rendering amount of time played on a specific save wasn't working, so I merged the two.  I think it mainly has to do with making calls to loadfromrenderedtext, but at least it's working now.  still going to require some debugging I'm sure, but in some cases it looks good.