modified buttons, savegame, timer, etc
moved a lot of the button stuff out to pregameui and stage classes, I feel like it's a bit simpler now.  been debugging savegame played time loading and saving correctly.  still a lot of debugging to do as far as displaying the played time on the new/load game scenes, also need to fix textures loading if player leaves gamestate 5 and wants to go back in the same session.