debugging again
I think the text textures are fixed now for the newgame/loadgame screens.  my friend the ghost characters from the saveDate looks like it may be there when player has created a new game, but gets stripped out after the game is loaded and then saved then reloaded, so fix that.  still trying to figure out where the problem is with the playedTime calculating incorrectly.  near as I can figure, the timer isn't pausing between saving and loading and is causing numbers that are higher than they're supposed to be.  it's a convoluted mess and it's going to take a while to clean up.