debugging
the amount of time the player has been in game at the newgame/loadgame screens looks to be reporting correctly now.  the ticks weren't calculating correctly, right now it works, but the code needs to be cleaned up for understandability/readability.  more testing needs to be done, but from the little bit that I've done on the fix it looks good so far.