new assets
upgraded the art work for habitation scenes as well as external view of the ship.  changed the x/y coords for the stations to coincide with designated locations in the art.  also set the save & quit button to render after the stations are rendered that way its always above the stations.