debugging
timer wasn't actually recording playedTime correctly before, now it looks like it is.  only tested it a moderate amount, but at least for the first savegame it seems to be working right.  the timer didn't seem to be restarting when it was necessary.  might not need to subtract pausedTicks from getTicks in the updatePlayedTime function in the timer class.