debugging
continuing to test for and fix bugs.  this update fixes an issue with loading an empty game.  before it would have very irregular, unintentional behaviour, now it doesn't load anything.  also the text textures for save game last played date and amount played on the newgame/loadgame screens is now being freed whenever the user starts/loads a game or goes back to main.  still need a 'really delete?' button for users that create a new game in a slot that has savegame data.  also need a delete savegame button on the newgame /loadgame screens that calls that prompt when the player wants to manually delete a save.