started on UI
Created a UI image, save & exit button is going to have to be changed to something like user presses 'esc' and a menu comes up or something, but yeah, for now just moved it to render below the UI image (technically it's slightly in front of UI image, but those pixels are alpha 0).  Not sure if I want ship details to show up until player upgrades the ship, but they are showing up right now, and of those, not sure if I want to track temperature.  pretty sure I'm going to just display percentages for most of the player levels/ship levels.  temp would obviously be celsius.  center area of UI I was thinking would be overarching things like time/date (for character), and money.  Date could start out as Sol 0.