planter interaction
planter interaction is now loading the buttons for the choices when interacting with the planter at the correct time and freeing them when player is not interacting with the planter.  unavailable choices are now rendered black and available choices are rendered white.  the textures for sown and sown & watered are now showing up correctly.