infirmary added
created art for infirmary as well as basic framework of the code for the infirmary.  it works like the other stations for the most part.  I could probably reduce the size (in pixels) of the image, but it works pretty well, even displays 'healthy' when user attempts to interact with it (so far).  eventually it'll be used to improve player health, probably using some kind of resource.