added more savegame datapoints
added some variables yesterday that weren't being saved and it was causing anomalous behaviour with the planter.  they're now being saved and loaded and the planter is rendering correctly, but these variables needed to be saved by the game anyway as they will be needed for the functionality of the planter.