walk animations, planter debugging
fixed some bugs with the planter, after testing it seems to work the way I want it to now.  the vegatative, unwatered planter image I guess never got added to the sprites folder, so I redid it.  I might reduce the cost of certain interactions with stations in the interest of balance.  added some of the animation frames for the character's space outfit.  so far up and down are working decent enough.