engineering module
I've been working on the engineering module code, added a couple of bg images that are basically the same as the hab ones except they say engineering in the bottom left instead of habitation, added a couple variables engInternalY1 & Y2 to be saved when game is saved.  inHab and inEng Bools will need to be saved and loaded whenever I add a mechanic to actually enter the hab section.  renderStage1 function will need to be updated a lotish to render the engineering stations as well as the hab stations are currently being rendered.