more engineering section work
I've been trying to set up a way to enter the engineering section, it's not quite there yet, but I know that I'm close.  created a small (about 10x10 px) station on the right side of the hab that is interactable, but it's not currently changing up the backgrounds hiding the hab stations, but we'll get there.  there will eventually be a short transition screen showing the player floating through the connecting tube between the bulkheads of the two modules.