module class
changing the structure of the game a little.  instead of everything loading and running from stage class I'm aiming for a heirarchy like stage>ship>module>station.  I made some of the starting changes, but it's going to take time to make this work correctly, when it does though, it'll be a lot prettier on the programming side.