problem solved
ship needed to be passed by reference to the player.move function, that's what was causing the ship gauges to not show up, as well as the stationTextures to not render.  this is probably what was going wrong with save/load game, might have fixed some of that stuff as well. Yay!