bug fixes, more opening sequence stuff
fixed the issue where the parallax background in the main stage of the game was slowing everything else down (specifically the player's walking / animation).  added text texture to encourage the player to open the newspaper in the opening sequence.  started a timed render function in texture.cpp, it's not ready.