update to research station
single bed button in research screen is now green to differentiate it from the description.  description is a little long, so trim it down.  scrap is not updating on the inventory screen, probably need to add an update inventory function.  researching single bed passes 2 hours, and costs some points of player needs.  the texture for the bed is working, bed needs to actually be more efficient now.