text texture rendering
I've had this theory that using for loops to render my arrays of text textures has been causing problems, so I've pulled the for loops and had things render one by one.  it seems to be working, but I noticed the 'sleep' button had an issue where it was showing 'upgrade' (though still responding like the sleep button).  it'll require further debugging to figure out what is causing this.